refactor: reorganize components & UI fiixes
