--- src/components/payment-summary/PaymentSummary.tsx
@@ -1,8 +1,0 @@
-import PaymentCard from "../payment/payment-card/PaymentCard";
-
-export const PaymentSummary = () => (
-  <div>
-    <span className="text-lg font-medium">Payment Summary</span>
-    <PaymentCard />
-  </div>
-);
